SUBMARINE-1209. Move folder and change dependency
