+++ submarine-server/server-submitter/submarine-k8s-agent/src/main/java/org/apache/submarine/server/k8s/agent/handler/NotebookHandler.java
@@ -28,8 +28,8 @@ import io.kubernetes.client.util.generic.options.ListOptions;
 import org.apache.submarine.server.api.common.CustomResourceType;
 import org.apache.submarine.server.api.notebook.Notebook;
 import org.apache.submarine.server.k8s.agent.util.RestClient;
-import org.apache.submarine.server.submitter.k8s.model.NotebookCR;
-import org.apache.submarine.server.submitter.k8s.model.NotebookCRList;
+import org.apache.submarine.server.submitter.k8s.model.notebook.NotebookCR;
+import org.apache.submarine.server.submitter.k8s.model.notebook.NotebookCRList;
 import org.apache.submarine.server.submitter.k8s.util.NotebookUtils;
 import io.kubernetes.client.util.generic.GenericKubernetesApi;
 import org.slf4j.Logger;
