+++ submarine-server/server-submitter/submitter-k8s/src/main/java/org/apache/submarine/server/submitter/k8s/model/notebook/NotebookStatus.java
@@ -16,26 +16,36 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-
-package org.apache.submarine.server.submitter.k8s.model;
-
-import java.util.Map;
+package org.apache.submarine.server.submitter.k8s.model.notebook;
 
 import com.google.gson.annotations.SerializedName;
+import io.kubernetes.client.openapi.models.V1ContainerState;
 
-public abstract class MLJobSpec {
-  
-  @SerializedName("backoffLimit")
-  private Integer backoffLimit = 3;
-  
-  public abstract Map<MLJobReplicaType, MLJobReplicaSpec> getReplicaSpecs();
-  public abstract void setReplicaSpecs(Map<MLJobReplicaType, MLJobReplicaSpec> replicaSpecs);
-  public Integer getBackoffLimit() {
-    return backoffLimit;
-  }
-  
-  public void setBackoffLimit(Integer backoffLimit) {
-    this.backoffLimit = backoffLimit;
+import java.util.List;
+
+public class NotebookStatus {
+
+  @SerializedName("conditions")
+  private List<NotebookCondition> conditions;
+
+  @SerializedName("readyReplicas")
+  private int readyReplicas;
+
+  @SerializedName("containerState")
+  private V1ContainerState containerState;
+
+  NotebookStatus() {
   }
 
+  public int getReadyReplicas() {
+    return readyReplicas;
+  }
+
+  public V1ContainerState getContainerState() {
+    return containerState;
+  }
+
+  public List<NotebookCondition> getConditions() {
+    return conditions;
+  }
 }
