+++ submarine-server/server-submitter/submitter-k8s/src/main/java/org/apache/submarine/server/submitter/k8s/model/tfjob/TFJob.java
@@ -23,7 +23,7 @@ import com.google.gson.annotations.SerializedName;
 
 import io.kubernetes.client.common.KubernetesObject;
 
-import org.apache.submarine.server.submitter.k8s.model.MLJob;
+import org.apache.submarine.server.submitter.k8s.model.mljob.MLJob;
 
 /**
  * It's the tf-operator's entry model.
