+++ submarine-server/server-submitter/submitter-k8s/src/main/java/org/apache/submarine/server/submitter/k8s/model/tfjob/TFJobReplicaType.java
@@ -20,7 +20,7 @@
 package org.apache.submarine.server.submitter.k8s.model.tfjob;
 
 import com.google.gson.annotations.SerializedName;
-import org.apache.submarine.server.submitter.k8s.model.MLJobReplicaType;
+import org.apache.submarine.server.submitter.k8s.model.mljob.MLJobReplicaType;
 
 public enum TFJobReplicaType implements MLJobReplicaType {
